+++ src/main/java/com/storage/site/model/rowmapper/UnitRowMapper.java
@@ -13,11 +13,11 @@ public class UnitRowMapper implements RowMapper<Unit> {
     @Override
     public Unit mapRow(ResultSet rs, int rowNumber) throws SQLException {
         return new Unit(
-                rs.getByte("id"),
+                rs.getInt("id"),
                 rs.getBoolean("is_large"),
                 rs.getBoolean("is_occupied"),
                 rs.getBoolean("is_delinquent"),
-                rs.getShort("days_delinquent"),
+                rs.getInt("days_delinquent"),
                 rs.getDate("start_date")
         );
     }
